requirements/base.txt
Django==1.11
psycopg2==2.7.3.2
Pillow==4.3.0
django-image-cropping==1.1.0
easy-thumbnails==2.5
